Fix message ambiguity and allow omitting pipe
The message shown on Special:DeleteBatch is commonly interpreted
as that the reason is optional, including the pipe that separates
the page name from the reason. While this is not necessarily an
issue, the real problem is that the script assumes there will be
a pipe and does not verify if the given line actually contained
one. This results in an error, but the script still continues to
run. Since the pipe is superfluous, allow omitting it.

Bug: T182455
Change-Id: If4d32cff2b09c9830ab07d17b93485d0b9be6cc8