Allow users to set a default deletion reason (e.g. "Spam", "Vandalism", etc.)
The default deletion reason will be used for all pages to be deleted unless explicitly overridden by specifying a different reason.

So suppose the default deletion reason is "Test page" and the input box looks like this:

Foo|Junk page
Bar
Baz
FooBaz|Created accidentally

The page "Foo" will be deleted with the reason "Junk page"; the page FooBaz will be deleted with the reason "Created accidentally".
The pages "Bar" and "Baz" will be deleted with the reason "Test page".

Change-Id: I4d9e5212f24994c72b7b9cb2174fc5d6e8a3a6a5